Upload of refactor so far to switch machines.
