+++ include/builtins.h
@@ -1,0 +1,7 @@
+#ifndef BUILTINS_H
+#define BUILTINS_H
+
+#define NUM_OF_BLTNS 5
+#define BNAME_BUFFER 8
+
+#endif
