+++ include/my_strings.h
@@ -1,0 +1,11 @@
+#ifndef MY_STRINGS_H
+#define MY_STRINGS_H
+
+#include <stdio.h>
+#include <stdlib.h>
+
+int string_compare(char *str1, char *str2);
+int string_length(char *str);
+char *copy_string(char *string);
+
+#endif
